+++ sites/svelte.dev/src/lib/server/docs/types.d.ts
@@ -21,11 +21,3 @@ export type Page = {
 	content: string;
 	sections: Section[];
 };
-
-export interface Type {
-	name: string;
-	comment: string;
-	snippet: string;
-	bullets: string[];
-	children: Type[];
-}
