+++ sites/svelte.dev/src/lib/server/markdown/types.d.ts
@@ -1,0 +1,7 @@
+export interface Type {
+	name: string;
+	comment: string;
+	snippet: string;
+	bullets: string[];
+	children: Type[];
+}
